Model: factor `Register.cpp` out of `Binary.cpp`
